# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gallery API",
    "description": "API for managing images in a gallery",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/gallery": {
      "post": {
        "summary": "Create a new image",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Image created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "summary": "Get all images",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "link": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gallery/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer"
          }
        }
      ],
      "put": {
        "summary": "Update an image",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image updated"
          },
          "404": {
            "description": "Image not found"
          }
        }
      },
      "delete": {
        "summary": "Delete an image",
        "responses": {
          "200": {
            "description": "Image deleted"
          },
          "404": {
            "description": "Image not found"
          }
        }
      }
    }
  }
}